chore: rename useFps (#907)
