+++ packages/core/index.ts
@@ -31,9 +31,9 @@ export * from './useElementVisibility'
 export * from './useEventBus'
 export * from './useEventListener'
 export * from './useEventSource'
-export * from './useFPS'
 export * from './useFavicon'
 export * from './useFetch'
+export * from './useFps'
 export * from './useFullscreen'
 export * from './useGeolocation'
 export * from './useIdle'
