+++ packages/core/useFps/demo.vue
