+++ packages/core/useFps/index.ts
